[doctor] fail on cmd.exe, launch expo install with powershell on windows
